Fix build break and extract kusto query modifying logic into helper class. Also this will help us create a unit test for it to make sure that is not broken.
