+++ src/Diagnostics.DataProviders/DataProviders/KustoDataProvider.cs
@@ -6,6 +6,7 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using Diagnostics.DataProviders.Interfaces;
+using Diagnostics.DataProviders.Utility;
 using Diagnostics.ModelsAndUtils.Models;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 
@@ -49,14 +50,14 @@ namespace Diagnostics.DataProviders
         public async Task<DataTable> ExecuteClusterQuery(string query, string cluster, string databaseName, string requestId, string operationName)
         {
             await AddQueryInformationToMetadata(query, cluster, operationName);
-            return await _kustoClient.ExecuteQueryAsync(MakeQueryCloudAgnostic(query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(databaseName) ?? databaseName, requestId, operationName);
+            return await _kustoClient.ExecuteQueryAsync(Helpers.MakeQueryCloudAgnostic(_kustoMap, query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(databaseName) ?? databaseName, requestId, operationName);
         }
 
         public async Task<DataTable> ExecuteQuery(string query, string stampName, string requestId = null, string operationName = null)
         {
             var cluster = await GetClusterNameFromStamp(stampName);
             await AddQueryInformationToMetadata(query, cluster, operationName);
-            return await _kustoClient.ExecuteQueryAsync(MakeQueryCloudAgnostic(query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, requestId, operationName);
+            return await _kustoClient.ExecuteQueryAsync(Helpers.MakeQueryCloudAgnostic(_kustoMap, query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, requestId, operationName);
         }
 
         public Task<KustoQuery> GetKustoClusterQuery(string query)
@@ -72,30 +73,8 @@ namespace Diagnostics.DataProviders
         public async Task<KustoQuery> GetKustoQuery(string query, string stampName, string operationName)
         {
             var cluster = await GetClusterNameFromStamp(stampName);
-            var kustoQuery = await _kustoClient.GetKustoQueryAsync(MakeQueryCloudAgnostic(query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, operationName);
+            var kustoQuery = await _kustoClient.GetKustoQueryAsync(Helpers.MakeQueryCloudAgnostic(_kustoMap, query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, operationName);
             return kustoQuery;
-        }
-
-        private string MakeQueryCloudAgnostic(string query)
-        {
-            var matches = Regex.Matches(query, @"cluster\((?<cluster>(.+))\).database\((?<database>(.+))\)\.");
-
-            if (matches.Any())
-            {
-                foreach (Match element in matches)
-                {
-                    var targetCluster = _kustoMap.MapCluster(element.Groups["cluster"].Value.Trim(new char[] { '\'', '\"' }));
-                    var targetDatabase = _kustoMap.MapCluster(element.Groups["database"].Value.Trim(new char[] { '\'', '\"' }));
-
-                    if (!string.IsNullOrWhiteSpace(targetCluster) && !string.IsNullOrWhiteSpace(targetDatabase))
-                    {
-                        query = query.Replace($"cluster({element.Groups["cluster"].Value})", $"cluster('{}')");
-                        query = query.Replace($"database({element.Groups["database"].Value})", $"database('{}')");
-                    }
-                }
-            }
-
-            return query;
         }
 
         internal async Task<DataTable> ExecuteQueryForHeartbeat(string query, string cluster, int timeoutSeconds, string requestId = null, string operationName = null)
@@ -110,7 +89,7 @@ namespace Diagnostics.DataProviders
 
         private async Task AddQueryInformationToMetadata(string query, string cluster, string operationName = null)
         {
-            var kustoQuery = await _kustoClient.GetKustoQueryAsync(MakeQueryCloudAgnostic(query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, operationName);
+            var kustoQuery = await _kustoClient.GetKustoQueryAsync(Helpers.MakeQueryCloudAgnostic(_kustoMap, query), _kustoMap.MapCluster(cluster) ?? cluster, _kustoMap.MapDatabase(_configuration.DBName) ?? _configuration.DBName, operationName);
             bool queryExists = false;
 
             queryExists = Metadata.PropertyBag.Any(x => x.Key == "Query" &&
