+++ src/Diagnostics.DataProviders/Utility/Helpers.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
+using Diagnostics.DataProviders.Interfaces;
 
 namespace Diagnostics.DataProviders.Utility
 {
@@ -18,6 +21,28 @@ namespace Diagnostics.DataProviders.Utility
                 default:
                     return "Secrets:DevKeyVaultName";
             }
+        }
+
+        public static string MakeQueryCloudAgnostic(IKustoMap kustoMap, string query)
+        {
+            var matches = Regex.Matches(query, @"cluster\((?<cluster>(.+))\).database\((?<database>(.+))\)\.");
+
+            if (matches.Any())
+            {
+                foreach (Match element in matches)
+                {
+                    var targetCluster = kustoMap.MapCluster(element.Groups["cluster"].Value.Trim(new char[] { '\'', '\"' }));
+                    var targetDatabase = kustoMap.MapCluster(element.Groups["database"].Value.Trim(new char[] { '\'', '\"' }));
+
+                    if (!string.IsNullOrWhiteSpace(targetCluster) && !string.IsNullOrWhiteSpace(targetDatabase))
+                    {
+                        query = query.Replace($"cluster({element.Groups["cluster"].Value})", $"cluster('{targetCluster}')");
+                        query = query.Replace($"database({element.Groups["database"].Value})", $"database('{targetDatabase}')");
+                    }
+                }
+            }
+
+            return query;
         }
     }
 }
